requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
executing==2.1.0
fonttools==4.54.1
frozenlist==1.5.0
gotrue==2.9.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
jiter==0.6.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.1.2
openai==1.52.2
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
postgrest==0.17.2
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
realtime==2.0.6
regex==2024.9.11
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
storage3==0.8.2
supabase==2.9.1
supafunc==0.6.2
tiktoken==0.8.0
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
websockets==13.1
wheel==0.44.0
yarl==1.17.0
